+++ src/Accounts/Authenticators/SilentAuthenticator.cs
@@ -71,7 +71,6 @@ namespace Microsoft.Azure.PowerShell.Authenticators
             SharedTokenCacheCredentialOptions options = AzConfigReader.IsWamEnabled(authority)
                 ? new SharedTokenCacheCredentialBrokerOptions(tokenCacheProvider.GetTokenCachePersistenceOptions())
                 : new SharedTokenCacheCredentialOptions(tokenCacheProvider.GetTokenCachePersistenceOptions());
-#pragma warning restore CS0618
             options.EnableGuestTenantAuthentication = true;
             options.ClientId = Constants.PowerShellClientId;
             options.Username = silentParameters.UserId;
@@ -85,6 +84,7 @@ namespace Microsoft.Azure.PowerShell.Authenticators
             }
             return options;
         }
+#pragma warning restore CS0618
 
         public override bool CanAuthenticate(AuthenticationParameters parameters)
         {